Adding tests for the new boolean buttons
Buttons as preset values is pretty useless. We would like to know when the
button has been pressed. Buttons should now behave more like checkboxes. They
have a value of their label when pressed and a value of '' when they have not
been pressed.

This adds failing tests to make that happen.