Gather features in get_all_features() with inheritance in mind
Previously, when a class inherited from another and both had checkers or
cleaners built using has_checker or has_cleaner, only a single class in the
entire ISA would be pulled in because one would clobber the rest. (Worse, it was
the one highest in the inheritance hierarchy.)

Now, we alias the feature names with the class name that uses it to avoid this.